+++ src/main/java/com/qcloud/Module/Emr.java
@@ -1,0 +1,15 @@
+package com.qcloud.Module;
+
+/*
+ * Tencent Cloud Elastic Map Reduce.
+ * <p>
+ * see API document: <a
+ * href="https://cloud.tencent.com/document/product/589/10214">
+ * https://cloud.tencent.com/document/product/589/10214</a>
+ * </p>
+ */
+public class Emr extends Base {
+    public Emr(){
+        serverHost = "emr.api.qcloud.com";
+    }
+}