+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ CHANGELOG
 latest (now)
 ============
 
+* add emr module
 * add athena module
 
 2.0.5 (2017-01-22)